Split out indigo_api.models.py into separate files
