+++ indigo_api/models/__init__.py
@@ -1,0 +1,4 @@
+from .places import *
+from .works import *
+from .documents import *
+from .tasks import *
